requirements.txt
faster-whisper
feedparser
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
opencc-python-reimplemented
tqdm
